+++ src/utils/createLayerAddToTree.js
@@ -1,30 +1,27 @@
 import {getLayerWhere} from "@masterportal/masterportalapi/src/rawLayerList";
 
 /**
- * Creates a layer containing the given features. If tree in menu shows folders, a new folder is created.
- * The name of the folder and the layer can be defined in config.json's property 'treeHighlightedFeatures'.
+ * Creates a layer containing the given features and shows it in menu tree.
  *
- * @param {Array} layerIds contains the ids of the layers, the features are got from
+ * @param {String} layerId contains the id of the layer, the features are got from
  * @param {Array} features contains the features to add to new layer
  * @param {String} treeType the treeType
  * @param {Object} thfConfig content of config.json's property 'treeHighlightedFeatures'
- * @param {String} callerKey i18n-key of thee caller of this function
  * @returns {void}
  */
-function createLayerAddToTree (layerIds, features, treeType, thfConfig = {}, callerKey = "") {
-    if (Array.isArray(layerIds)) {
-        const layerId = layerIds.join("_"),
-            layerNameKey = thfConfig.layerName ? thfConfig.layerName : "common:tree.selectedFeatures",
-            layerName = i18next.t(layerNameKey) + " (" + i18next.t(callerKey) + ")",
-            originalLayer = getLayer(layerIds[0]);
+function createLayerAddToTree (layerId, features, treeType, thfConfig = {}) {
+    if (layerId) {
+        const layerNameKey = thfConfig.layerName ? thfConfig.layerName : "common:tree.selectedFeatures",
+            originalLayer = getLayer(layerId);
 
         if (originalLayer) {
-            const id = layerId + "_" + callerKey,
-                attributes = setAttributes(originalLayer, id, layerName, layerNameKey, callerKey, treeType);
-            let highlightLayer = Radio.request("ModelList", "getModelByAttributes", {"name": layerName}),
+            const originalLayerName = originalLayer.get("name").replace(i18next.t(layerNameKey), "").trim(),
+                layerName = i18next.t(layerNameKey) + " " + originalLayerName,
+                id = layerId.indexOf(originalLayerName) === -1 ? layerId + "_" + originalLayerName : layerId,
+                attributes = setAttributes(originalLayer, id, layerName, layerNameKey, treeType);
+            let highlightLayer = Radio.request("ModelList", "getModelByAttributes", {"id": id}),
                 layerSource = null;
 
-            createFolderIfNotExists("markedDataFolder", treeType, thfConfig);
             if (!highlightLayer) {
                 highlightLayer = addLayerModel(attributes, id);
             }
@@ -36,7 +33,7 @@ function createLayerAddToTree (layerIds, features, treeType, thfConfig = {}, cal
             refreshTree(treeType);
         }
         else {
-            console.warn("Layer with id ", layerIds[0], " not found, folder or layer \"", layerName + "\" was not created!");
+            console.warn("Layer with id ", layerId, " not found, with features was not created!");
         }
     }
 }
@@ -84,32 +81,15 @@ function refreshTree (treeType) {
 }
 
 /**
- * Creates a new Folder in menu tree, if not exists.
- * @param {String} id of the layer
- * @param {String} treeType the treeType
- * @param {Object} thfConfig content of config.json's property 'treeHighlightedFeatures'
- * @returns {void}
- */
-function createFolderIfNotExists (id, treeType, thfConfig) {
-    if ((treeType === "custom" || treeType === "default") && Radio.request("Parser", "getItemByAttributes", {id: id}) === undefined) {
-        const folderNameKey = thfConfig.folderName ? thfConfig.folderName : "common:tree.selectedData";
-
-        Radio.trigger("Parser", "addFolder", i18next.t(folderNameKey), id, "tree", 0, true, folderNameKey, false, false, "SelectedLayer");
-        Radio.trigger("ModelList", "renderTree");
-    }
-}
-
-/**
  * Copies the attributesof the given layer and adapts them.
  * @param {Object} layer to copy attributes of
  *  @param {String} id of the layer
  * @param {String} layerName name of the layer
  * @param {String} layerNameKey i18n-key of the name of the layer
- * @param {String} callerKey i18n-key of thee caller of the function 'createLayerAddToTree'
  * @param {String} treeType the treeType
  * @returns {Object} the adapted attributes
  */
-function setAttributes (layer, id, layerName, layerNameKey, callerKey, treeType) {
+function setAttributes (layer, id, layerName, layerNameKey, treeType) {
     const attributes = {...layer.attributes};
 
     // initial set selected and visibility to false, else source is updated before features are added
@@ -120,10 +100,10 @@ function setAttributes (layer, id, layerName, layerNameKey, callerKey, treeType)
     attributes.alwaysOnTop = true;
     attributes.name = layerName;
     attributes.selectionIDX = 1000;
-    attributes.parentId = treeType === "custom" || treeType === "default" ? "markedDataFolder" : "tree";
+    attributes.parentId = treeType === "custom" || treeType === "default" ? "SelectedLayer" : "tree";
     attributes.i18nextTranslate = (setter) => {
         if (typeof setter === "function" && i18next.exists(layerNameKey)) {
-            setter("name", i18next.t(layerNameKey) + " (" + i18next.t(callerKey) + ")");
+            setter("name", i18next.t(layerNameKey) + " " + attributes.name);
         }
     };